fix: move MetaEvaluatorState and DataMetaData to their respective serialization.py files
