+++ src/meta_evaluator/meta_evaluator/metaevaluator.py
@@ -14,12 +14,11 @@ from ..llm_client.LLM_client import LLMClient
 from ..llm_client.openai_client import OpenAIClient, OpenAIConfig
 from ..llm_client.azureopenai_client import AzureOpenAIClient, AzureOpenAIConfig
 from ..llm_client.serialization import (
-    MetaEvaluatorState,
-    DataMetadata,
     OpenAISerializedState,
     AzureOpenAISerializedState,
 )
 from ..data.EvalData import EvalData, SampleEvalData
+from ..data.serialization import DataMetadata
 from .exceptions import (
     MissingConfigurationException,
     ClientAlreadyExistsException,
@@ -27,6 +26,7 @@ from .exceptions import (
     DataAlreadyExistsException,
     DataFilenameExtensionMismatchException,
 )
+from .serialization import MetaEvaluatorState
 
 # Error message constants
 INVALID_JSON_STRUCTURE_MSG = "Invalid JSON structure in state file"
